requirements.txt
python-telegram-bot==20.0
openpyxl
python-dotenv
pytz
